Исправили косячки
